Continue moving files
