+++ Guardian.Backend/Guardian.Domain/ApiException.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Globalization;
+
+namespace Guardian.Domain
+{
+    public class ApiException : Exception
+    {
+        public ApiException() : base() { }
+
+        public ApiException(string message) : base(message) { }
+
+        public ApiException(string message, params object[] args)
+            : base(string.Format(CultureInfo.CurrentCulture, message, args))
+        {
+        }
+    }
+}
