--- Guardian.Backend/Guardian.Infrastructure/Extension/ConfigureContainer.cs
@@ -1,24 +1,0 @@
-﻿using Guardian.Infrastructure.Middleware;
-using Microsoft.AspNetCore.Builder;
-
-namespace Guardian.Infrastructure.Extension
-{
-    public static class ConfigureContainer
-    {
-        public static void ConfigureCustomExceptionMiddleware(this IApplicationBuilder app)
-        {
-            app.UseMiddleware<CustomExceptionMiddleware>();
-        }
-
-        public static void ConfigureSwagger(this IApplicationBuilder app)
-        {
-            app.UseSwagger();
-
-            app.UseSwaggerUI(setupAction =>
-            {
-                setupAction.SwaggerEndpoint("/swagger/OpenAPISpecification/swagger.json", "Onion Architecture API");
-                setupAction.RoutePrefix = "OpenAPI";
-            });
-        }
-    }
-}
