+++ Guardian.Backend/Guardian.Infrastructure/Guardian.Infrastructure.Email/IEmailService.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading.Tasks;
+using Guardian.Domain.Settings;
+
+namespace Guardian.Infrastructure.Email
+{
+    public interface IEmailService
+    {
+        Task SendEmailAsync(MailRequest mailRequest);
+
+    }
+}
