--- Guardian.Backend/Guardian.Infrastructure/Mapping/CustomerProfile.cs
@@ -1,17 +1,0 @@
-﻿using AutoMapper;
-using Guardian.Domain.Entities;
-using Guardian.Infrastructure.ViewModel;
-
-namespace Guardian.Infrastructure.Mapping
-{
-    public class CustomerProfile : Profile
-    {
-        public CustomerProfile()
-        {
-            CreateMap<CustomerModel, Customer>()
-                .ForMember(dest => dest.Id,
-                        opt => opt.MapFrom(src => src.CustomerId))
-                .ReverseMap();
-        }
-    }
-}
