--- Guardian.Backend/Guardian.Infrastructure/ViewModel/CustomerModel.cs
@@ -1,18 +1,0 @@
-﻿
-namespace Guardian.Infrastructure.ViewModel
-{
-    public class CustomerModel
-    {
-        public int CustomerId { get; set; }
-        public string CustomerName { get; set; }
-        public string ContactName { get; set; }
-        public string ContactTitle { get; set; }
-        public string Address { get; set; }
-        public string City { get; set; }
-        public string Region { get; set; }
-        public string PostalCode { get; set; }
-        public string Country { get; set; }
-        public string Phone { get; set; }
-        public string Fax { get; set; }
-    }
-}
