--- Guardian.Backend/Guardian.Persistence/ApplicationDbContext.cs
@@ -1,45 +1,0 @@
-﻿using Guardian.Domain.Entities;
-using Microsoft.EntityFrameworkCore;
-using System.Threading.Tasks;
-
-namespace Guardian.Persistence
-{
-    public class ApplicationDbContext : DbContext, IApplicationDbContext
-    {
-        // This constructor is used of runit testing
-        public ApplicationDbContext()
-        {
-
-        }
-
-        public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options) : base(options)
-        {
-            ChangeTracker.QueryTrackingBehavior = QueryTrackingBehavior.NoTracking;
-        }
-
-        public DbSet<Customer> Customers { get; set; }
-        public DbSet<Order> Orders { get; set; }
-        public DbSet<Product> Products { get; set; }
-        public DbSet<Category> Categories { get; set; }
-        public DbSet<Supplier> Suppliers { get; set; }
-
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            modelBuilder.Entity<OrderDetail>().HasKey(o => new { o.OrderId, o.ProductId });
-        }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            if (!optionsBuilder.IsConfigured)
-            {
-                optionsBuilder
-                .UseSqlServer("DataSource=app.db");
-            }
-        }
-
-        public async Task<int> SaveChangesAsync()
-        {
-            return await base.SaveChangesAsync();
-        }
-    }
-}
