--- Guardian.Backend/Guardian.Persistence/IApplicationDbContext.cs
@@ -1,17 +1,0 @@
-﻿using Guardian.Domain.Entities;
-using Microsoft.EntityFrameworkCore;
-using System.Threading.Tasks;
-
-namespace Guardian.Persistence
-{
-    public interface IApplicationDbContext
-    {
-        DbSet<Category> Categories { get; set; }
-        DbSet<Customer> Customers { get; set; }
-        DbSet<Order> Orders { get; set; }
-        DbSet<Product> Products { get; set; }
-        DbSet<Supplier> Suppliers { get; set; }
-
-        Task<int> SaveChangesAsync();
-    }
-}
