+++ Guardian.Backend/Guardian.Service/Contract/IAuthenticatedUserService.cs
@@ -1,5 +1,4 @@
-﻿
-namespace Guardian.Service.Contract
+﻿namespace Guardian.Service.Contract
 {
     public interface IAuthenticatedUserService
     {
