--- Guardian.Service.Contract/Guardian.Service.Contract.csproj
@@ -1,11 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Guardian.Backend\Guardian.Domain\Guardian.Domain.csproj" />
-  </ItemGroup>
-
-</Project>
