--- Guardian.Service.Identity/Guardian.Service.Identity.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Identity" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="5.0.17" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Guardian.Backend\Guardian.Domain\Guardian.Domain.csproj" />
-    <ProjectReference Include="..\Guardian.Service.Contract\Guardian.Service.Contract.csproj" />
-  </ItemGroup>
-
-</Project>
